Abstract out http PUT
